corrected quit menu option logic and adjusted confirm dialog text
the file > quit menu option now quits the application, instead of just closing the current window. The user will still be asked to confirm this before quitting.

The confirm dialogs have been adjusted, so that the user is warned if selecting file > quit that ALL editor windows with unsaved changes will close, vs only the selected window if closed by the window controls